Move self-force analytic data to self-force system directory
Consolidate the self-force code in one directory, so it's not
scattered across the code base. This has been an obstacle for
new people working with the code. Only the executable is still
in a separate directory tree.
