+++ src/Elliptic/Systems/SelfForce/GeneralRelativity/AnalyticData/CircularOrbit.hpp
