+++ src/Elliptic/Systems/SelfForce/Scalar/AnalyticData/CircularOrbit.hpp
