+++ src/Elliptic/Systems/SelfForce/Scalar/Events/ObserveSelfForce.hpp
@@ -21,6 +21,7 @@
 #include "Domain/ElementLogicalCoordinates.hpp"
 #include "Domain/Structure/ElementId.hpp"
 #include "Domain/Tags.hpp"
+#include "Elliptic/Systems/SelfForce/Scalar/AnalyticData/CircularOrbit.hpp"
 #include "Elliptic/Systems/SelfForce/Scalar/Tags.hpp"
 #include "Elliptic/Tags.hpp"
 #include "IO/Observer/GetSectionObservationKey.hpp"
@@ -39,8 +40,6 @@
 #include "Parallel/TypeTraits.hpp"
 #include "ParallelAlgorithms/Events/Tags.hpp"
 #include "ParallelAlgorithms/EventsAndTriggers/Event.hpp"
-#include "PointwiseFunctions/AnalyticData/SelfForce/Scalar/CircularOrbit.hpp"
-#include "PointwiseFunctions/InitialDataUtilities/Background.hpp"
 #include "Utilities/ErrorHandling/Assert.hpp"
 #include "Utilities/ErrorHandling/Error.hpp"
 #include "Utilities/Functional.hpp"
