+++ src/PointwiseFunctions/AnalyticData/CMakeLists.txt
@@ -24,5 +24,4 @@ add_subdirectory(RadiationTransport)
 add_subdirectory(NewtonianEuler)
 add_subdirectory(Punctures)
 add_subdirectory(ScalarTensor)
-add_subdirectory(SelfForce)
 add_subdirectory(Xcts)
