+++ tests/Unit/Elliptic/Systems/SelfForce/Scalar/CMakeLists.txt
@@ -2,5 +2,6 @@
 # See LICENSE.txt for details.
 
 add_subdirectory(AmrCriteria)
+add_subdirectory(AnalyticData)
 add_subdirectory(BoundaryConditions)
 add_subdirectory(Events)
