+++ tests/Unit/PointwiseFunctions/AnalyticData/CMakeLists.txt
@@ -27,5 +27,4 @@ add_subdirectory(NewtonianEuler)
 add_subdirectory(Punctures)
 add_subdirectory(RadiationTransport)
 add_subdirectory(ScalarTensor)
-add_subdirectory(SelfForce)
 add_subdirectory(Xcts)
