--- tests/Unit/PointwiseFunctions/AnalyticData/SelfForce/CMakeLists.txt
@@ -1,5 +1,0 @@
-# Distributed under the MIT License.
-# See LICENSE.txt for details.
-
-add_subdirectory(GeneralRelativity)
-add_subdirectory(Scalar)
